[Clang][Sema] wip: Introduce the scaffolding for the "explicit params" feature
This patch adds the basic ground-work for supporting an optional new
feature through which explicit (non-converting) parameters will be
possible to be added to functions. No changes were made to the actual
AST or the feature itself, only the outermost user-facing interface was
created accordingly.

A new flag, `-fexplicit-function-params` toggles the feature without
which the existing logic is kept, but an appropriate error explains that
explicit params are an extension. All other locations where `explicit`
is forbidden will not be affected.